chore: add eslint architecture guardrails
