+++ scripts/guards/no-direct-invoke.sh
@@ -1,6 +1,6 @@
 #!/usr/bin/env bash
 set -euo pipefail
-hits=$(git grep -n -F "invoke(" -- src | grep -v "src/api/call.ts" || true)
+hits=$(git grep -n -F "invoke(" -- src | grep -v "src/lib/ipc/call.ts" || true)
 if [[ -n "$hits" ]]; then
   echo "❌ Direct invoke() calls found:"
   echo "$hits"
