+++ src/lib/ipc/clipboard.ts
@@ -1,0 +1,1 @@
+export { writeText } from "@tauri-apps/plugin-clipboard-manager";
