+++ src/lib/ipc/core.ts
@@ -1,0 +1,1 @@
+export { convertFileSrc } from "@tauri-apps/api/core";
