+++ src/lib/ipc/dialog.ts
@@ -1,0 +1,2 @@
+export { open } from "@tauri-apps/plugin-dialog";
+export type { OpenDialogOptions, DialogFilter } from "@tauri-apps/plugin-dialog";
